Maintain existing API conventions
Before now the packages were split into SugarExt and SugarGestures,
the existing build didn't consider this and put everything into one,
which breaks the API as we'll have to track and change those uses.

This changes that and maintains the existing API convention so we don't
have to worry about changing existing use.

I've also added some of the other files that aren't dependent on any x11
related library as I'd like to test this.

I removed sugar-file-attributes entirely because that's something we never
want to do, journal objects exist for a reason and I'd like to keep the
current model without changing it.
This has been replaced by sugar-fatattr which was used earlier and I decided
that we should keep it based on its use and it doesn't break anything.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>