Change API version
The API version change is necessary so as not to confuse this with
the existing libraries.

I've also ported sugar-clipboard to Gtk4, and added it here because Sugar
still depends on it, and this is for my tests.

I've also added install_gir and install_typelib as this made
the install easier and installed the necessary files to the needed
location, so I wouldn't have to be bothered about them.
These files were generated but not installed in the right directory
before now, and this makes sure they are so Python can see them
on import.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>